Changed the activity management structure, added department activity backend
