+++ backend/kpi-tracker-models/src/main/java/org/pahappa/systems/kpiTracker/models/activities/DepartmentActivity.java
@@ -1,0 +1,124 @@
+package org.pahappa.systems.kpiTracker.models.activities;
+
+import org.pahappa.systems.kpiTracker.models.goals.DepartmentGoal;
+import org.pahappa.systems.kpiTracker.models.organization_structure.Department;
+import org.pahappa.systems.kpiTracker.models.systemSetup.enums.ActivityPriority;
+import org.pahappa.systems.kpiTracker.models.systemSetup.enums.ActivityStatus;
+import org.pahappa.systems.kpiTracker.models.systemSetup.enums.ActivityType;
+import org.sers.webutils.model.BaseEntity;
+
+import javax.persistence.*;
+import java.util.Date;
+
+@Entity
+@Table(name = "department_activities")
+public class DepartmentActivity extends BaseEntity {
+    private String title;
+    private String description;
+    private ActivityStatus status;
+    private ActivityPriority priority;
+    private ActivityType activityType;
+    private Date plannedStartDate;
+    private Date plannedEndDate;
+    private double targetValue;
+    private double actualValue;
+    private DepartmentGoal departmentGoal;
+    private Department department;
+
+    @Column(name = "title", nullable = false, length = 200)
+    public String getTitle() {
+        return title;
+    }
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    @Column(name = "description", columnDefinition = "TEXT")
+    public String getDescription() {
+        return description;
+    }
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "status", nullable = false)
+    public ActivityStatus getStatus() {
+        return status;
+    }
+    public void setStatus(ActivityStatus status) {
+        this.status = status;
+    }
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "priority", nullable = false)
+    public ActivityPriority getPriority() {
+        return priority;
+    }
+    public void setPriority(ActivityPriority priority) {
+        this.priority = priority;
+    }
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "activity_type", nullable = false)
+    public ActivityType getActivityType() {
+        return activityType;
+    }
+    public void setActivityType(ActivityType activityType) {
+        this.activityType = activityType;
+    }
+
+    @Temporal(TemporalType.DATE)
+    @Column(name = "planned_start_date", nullable = false)
+    public Date getPlannedStartDate() {
+        return plannedStartDate;
+    }
+    public void setPlannedStartDate(Date plannedStartDate) {
+        this.plannedStartDate = plannedStartDate;
+    }
+
+    @Temporal(TemporalType.DATE)
+    @Column(name = "planned_end_date", nullable = false)
+    public Date getPlannedEndDate() {
+        return plannedEndDate;
+    }
+    public void setPlannedEndDate(Date plannedEndDate) {
+        this.plannedEndDate = plannedEndDate;
+    }
+
+    @Column(name = "target_value", nullable = false)
+    public double getTargetValue() {
+        return targetValue;
+    }
+    public void setTargetValue(double targetValue) {
+        this.targetValue = targetValue;
+    }
+
+    @Column(name = "actual_value")
+    public double getActualValue() {
+        return actualValue;
+    }
+    public void setActualValue(double actualValue) {
+        this.actualValue = actualValue;
+    }
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "department_goal_id", nullable = true)
+    public DepartmentGoal getDepartmentGoal() {
+        return departmentGoal;
+    }
+
+    public void setDepartmentGoal(DepartmentGoal departmentGoal) {
+        this.departmentGoal = departmentGoal;
+    }
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "department_id", nullable = true)
+    public Department getDepartment() {
+        return department;
+    }
+
+    public void setDepartment(Department department) {
+        this.department = department;
+    }
+}
