+++ backend/kpi-tracker-services/src/main/java/org/pahappa/systems/kpiTracker/core/services/activities/DepartmentActivityService.java
@@ -1,0 +1,8 @@
+package org.pahappa.systems.kpiTracker.core.services.activities;
+
+import org.pahappa.systems.kpiTracker.core.services.GenericService;
+import org.pahappa.systems.kpiTracker.models.activities.DepartmentActivity;
+
+public interface DepartmentActivityService extends GenericService<DepartmentActivity> {
+    Object getObjectById(String id);
+}
