+++ backend/kpi-tracker-services/src/main/java/org/pahappa/systems/kpiTracker/core/services/activities/impl/DepartmentActivityServiceImpl.java
@@ -1,8 +1,8 @@
 package org.pahappa.systems.kpiTracker.core.services.activities.impl;
 
-import org.pahappa.systems.kpiTracker.core.services.activities.ActivityService;
+import org.pahappa.systems.kpiTracker.core.services.activities.DepartmentActivityService;
 import org.pahappa.systems.kpiTracker.core.services.impl.GenericServiceImpl;
-import org.pahappa.systems.kpiTracker.models.activities.Activity;
+import org.pahappa.systems.kpiTracker.models.activities.DepartmentActivity;
 import org.pahappa.systems.kpiTracker.utils.Validate;
 import org.sers.webutils.model.exception.OperationFailedException;
 import org.sers.webutils.model.exception.ValidationFailedException;
@@ -11,15 +11,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional
-public class ActivityServiceImpl extends GenericServiceImpl<Activity> implements ActivityService {
+public class DepartmentActivityServiceImpl extends GenericServiceImpl<DepartmentActivity> implements DepartmentActivityService {
     @Override
-    public Activity saveInstance(Activity entityInstance) throws ValidationFailedException, OperationFailedException {
+    public DepartmentActivity saveInstance(DepartmentActivity entityInstance) throws ValidationFailedException, OperationFailedException {
         Validate.notNull(entityInstance, "Missing details");
         return save(entityInstance);
     }
 
     @Override
-    public boolean isDeletable(Activity instance) throws OperationFailedException {
+    public boolean isDeletable(DepartmentActivity instance) throws OperationFailedException {
         return true;
     }
